requirements.txt
fastapi==0.95.1
sqlalchemy==2.0.21
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-jose==3.3.0
pydantic==1.10.9
passlib==1.7.4
python-multipart==0.0.6
alembic== 1.12.0